git-svn-id: file:///mnt/data/Workspace/zie87/game_development/tmp/hare-game-engine/p/trunk@200 67360d68-41fe-4630-88ca-a5c15c9c3197

+++ editor/ConfigManager.cpp
@@ -29,7 +29,7 @@ namespace hare
     {
         HARE_META_F(openFilePath, String, propReadOnly)
         HARE_META_F(openFileFilters, String, propReadOnly)
-        HARE_META_F(openFileFilterIndex, int, propHide)
+        HARE_META_F(openFileFilterIndex, int8, propHide)
         HARE_META_F(font, String, propReadOnly)
         HARE_META_F(defaultEncoding, String, propReadOnly)
         HARE_ENUM(show_folds, uint8, Bool)
